% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{control}
\alias{control}
\title{Control function for parameter and variance estimation}
\usage{
control(
  gtol = 1e-06,
  maxiter = 1000,
  eps = 1e-06,
  getBreslowJumps = TRUE,
  standardize = TRUE,
  cindex = FALSE,
  multicore = FALSE,
  mcores = detectCores() - 1
)
}
\arguments{
\item{gtol}{Numeric: specified for `fitter`=`crrc`; is the threshold for stopping the algorithm based on
a function of the gradient. The iteration stops when this function
of the gradient is less than `gtol`. Default is 1e-6.
algorithm. A value of 0 computes scores and variance at `init`,
but performs no iterations. Default is 10.}

\item{maxiter}{Integer: the maximum number of iterations to achieve convergence. Default is 1000.}

\item{eps}{Numeric: specified for `fitter`=`fastCrr`; the algorithm stops when the relative change in any
coefficient is less than `eps`. Default is 1E-6.}

\item{getBreslowJumps}{Logical: specified for `fitter`=`fastCrr`; whether to output jumps in Breslow estimator
for the cumulative hazard. Default is TRUE.}

\item{standardize}{Logical: specified for `fitter`=`fastCrr`; whether to standardize the design matrix.
Default is TRUE.}

\item{cindex}{Logical: whether to compute and output the concordance index for the model}

\item{multicore}{Logical: whether to use multicore processing. Default is FALSE.}

\item{mcores}{Integer: number of cores to use for multicore processing. Default is number of cores - 1.}
}
\value{
A list of control parameters configured as specified.
}
\description{
This function allows customization of various aspects of the `fitter` argument functions in `fastCrrC`
and also allows for multicore processing.
}
