% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastCrrC.R
\name{fastCrrC}
\alias{fastCrrC}
\title{Fast implementation of competing risks regression for clustered data}
\usage{
fastCrrC(
  formula,
  data,
  subset,
  cause,
  id,
  fitter = c("fastCrr", "crrc"),
  B = 0,
  bMeth = c("twoStep", "cluStep"),
  mcontrols = control(gtol = 1e-06, maxiter = 1000, eps = 1e-06, getBreslowJumps = TRUE,
    standardize = TRUE, cindex = FALSE, multicore = FALSE, mcores = detectCores() - 1)
)
}
\arguments{
\item{formula}{a formula expression, of the form `response ~ predictors`.
The `response` is a `Surv` object with right censoring.
In the case of competing risks, the competing risks indicator can be specified as the status indicator,
with 0 indicating right censoring, and 1, 2, ... indicating events other than right censoring.}

\item{data}{is an optional data.frame in which to interpret the variables occurring in the ‘formula’.}

\item{subset}{is an optional vector specifying a subset of observations to be used in the fitting process.}

\item{cause}{is the event of interest; default at 1 when not specified.}

\item{id}{is an optional vector used to identify the clsuters.
When missing, each individual row of `data` is presumed to represent a distinct subject.
The length of `id` should be the same as the number of observations.}

\item{fitter}{is an optional character string specifying the estimating procedures.
The available options are "fastCrr" and "crrc" corresponding to
the two-way linear scan with bootstrap for variance estimation and
the traditional marginal competing risk model using asymptotic variance estimation.
We recommend the two-way linear scan for n > 5000 or large sample or cluster and cluster sizes.}

\item{B}{is a numerical  value specifies the resampling number.
When B = 0, only the beta estimate will be displayed.}

\item{bMeth}{is an optional character string specifying the bootstrap method when `B > 0`.
Available methods are the clustered and two-step bootstrap.}

\item{mcontrols}{is an optional list with parameters that pass to fastCrr or crrc, such as maxiter, tolerance, and more.
The arguments `gtol`, `maxiter`, and `eps` can be used when `fitter`=`crrc`.
The arguments `getBreslowJumps`, `maxiter`, `standardize`, `multicore`, and `mcores` can be only be set when `fitter`=`fastCrr`.}

\item{cindex}{a logical value indicating whether the concordance index should be calculated.
Defaults to FALSE.}

\item{multicore}{a logical value indicating whether parallel processing should be used for computing the variance estimates.
Defaults to FALSE.}
}
\description{
Performs the marginal clustered competing risks regression based on the subdistribution hazards model.
}
\details{
This is based on the subdistribution hazards model using the two-way linear scan approach to estimate model parameters and bootstrap to compute the variances
}
\examples{
set.seed(1)
dat <- simCR(n = 10, al = .3, ga = .6, clsize = 2:10, summary = TRUE)

fit1 <- fastCrrC(Surv(time, status) ~ Z.1 + Z.2, data = dat, B = 10, fitter = "fastCrr")
fit2 <- fastCrrC(Surv(time, status) ~ Z.1 + Z.2, data = dat, B = 10, fitter = "crrc")

summary(fit1)
summary(fit2)

}
