% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fastCrrC}
\alias{plot.fastCrrC}
\title{Draw CIF}
\usage{
\method{plot}{fastCrrC}(x, type = c("one", "all", "stack"), newdata)
}
\arguments{
\item{x}{is an "\code{fastCrrC}" object}

\item{type}{types of plots to produce.
The available options are: one = plot CIF for the specified cause, all = plot CIF for all causes, stack = stack CIFs for all causes
\code{type = "all"} and \code{type = "stack"} will call \code{extend.fastCrrC()}}

\item{newdata}{is a data frame for an optional new data for CIF. Default to a vector of zero,
in which case the \code{plot()} function plots the baseline CIF.}
}
\description{
Draw CIF
}
\examples{
set.seed(1)
dat <- simCR(n = 100, al = .3, ga = .6, cen = .4, clsize = 2:5, summary = TRUE)
dat$status <- I(dat$status > 0)* sample(1:5, nrow(dat), TRUE)

foo <- fastCrrC(Surv(time, status) ~ Z.1 + Z.2, data = dat)
str(foo)
plot(foo)
plot(foo, newdata = data.frame(Z.1 = .5, Z.2 = 1))

## Plotting CIFs that involves other causes will automatically update the fit
plot(foo, type = "all")
plot(foo, type = "all", newdata = data.frame(Z.1 = .5, Z.2 = 1))
str(foo)

## Stacking CIF
plot(foo, type = "stack")
plot(foo, type = "stack", newdata = data.frame(Z.1 = .5, Z.2 = 1))
}
