% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCR.R
\name{simCR}
\alias{simCR}
\title{Function to generate simulated competing risk data}
\usage{
simCR(
  n = 100,
  al = 0.3,
  ga = 0.3,
  rho = 0,
  cen = 0.65,
  clsize = 2:50,
  summary = FALSE
)
}
\arguments{
\item{n}{is an integer specifying the number of clusters. Default at 100.}

\item{al}{is a numerical value generated from a positive stable distribution for cause 1 to control the within cluster correlation. Default is set at 0.3.}

\item{ga}{is a numerical value generated from a positive stable distribution for cause 2 to control the within cluster correlation. Default is set at 0.3.}

\item{rho}{is a numerical value specifying the degree of the within cluster dependence.}

\item{cen}{is a numerical value specifying the censoring rate. Only 0\%, 20\%, 40\%, 60\%, 65\%, and 80\% are currently allowed.}

\item{clsize}{an integer specifying the range of cluster sizes.
Each cluster size is randomly generated from a discrete uniform distribution with range [a, b].
where a and b are set to achieve the desired censoring rates of 0\%, 20\%, 40\%, 60\%, 65\%, and 80\%.}

\item{summary}{is an indicator indicating whether to print data summary}
}
\description{
The function \code{simCR()} generates data from a hypothetical multicenter.
To generate correlated failure times, the subdistribution hazard with positive frailty is used to
maintain the PH assumption on sub-distribution.
The dataframe returns the cluster IDs, event status, event time, censoring time
}
\examples{
set.seed(1)
head(simCR(n = 10, al = .3, ga = .6, summary = TRUE))
}
